requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
supabase
python-dotenv
